remove block factory file
